+++ src/components/editor/markdown-preview/replace-components/vimeo/vimeo-frame.tsx
@@ -2,7 +2,7 @@ import { DomElement } from 'domhandler'
 import React, { ReactElement, useCallback } from 'react'
 import { OneClickEmbedding } from '../one-click-frame/one-click-embedding'
 import '../one-click-frame/one-click-embedding.scss'
-import { testSingleVideoParagraph, VideoFrameProps } from '../video_util'
+import { testSingleVideoParagraph, VideoFrameProps } from '../video-util'
 
 const getElementReplacement = (node: DomElement, counterMap: Map<string, number>): (ReactElement | undefined) => {
   const videoId = testSingleVideoParagraph(node, 'vimeo')
